fixed bugs

+++ 0x1E-search_algorithms/1-binary.c
@@ -30,7 +30,7 @@ int binary_search(int *array, size_t size, int value)
 
 		if (array[mid] == value)
 		{
-			return (mid);
+			return ((int)mid);
 		}
 		else if (array[mid] < value)
 		{
